+++ src/test/components/Menu.spec.js
@@ -3,11 +3,11 @@
 import React from 'react'
 import { render } from '@testing-library/react'
 import Menu from '../../components/Menu'
-import { BrowserRouter } from 'react-router-dom'
+import { MemoryRouter } from 'react-router-dom'
 
 describe('Menu.js', () => {
   it('renders without crashing', () => {
-    const { getByTestId } = render(<BrowserRouter><Menu /></BrowserRouter>)
+    const { getByTestId } = render(<MemoryRouter><Menu /></MemoryRouter>)
     expect(getByTestId('menu')).toBeTruthy()
   })
 })