+++ CHANGELOG.md
@@ -1,8 +1,13 @@
-# ${project-name} Change Log
+# spotz Change Log
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [0.3.1] 2020-05-09
+### Added
+- Tests for all Routers
+- Tests for some Place Pages (need to add more)
 
 ## [0.3.0] 2020-05-08
 ### Fixed